+++ spec/cache_spec.rb
@@ -39,7 +39,7 @@ describe "Caching" do
     
     # force re read meta files creates all files if necessary
     FileUtils.rm_rf("spec/testdata/cache/main/.bake")
-    Bake.startBake("cache/main", ["-b", "test", "-v3", "--ignore_cache"])
+    Bake.startBake("cache/main", ["-b", "test", "-v3", "--ignore-cache"])
     expect($mystring.split("Loading and caching").length).to be == 7
     expect($mystring.split("Loading cached").length).to be == 3
     expect($mystring.split("Info: cache is up-to-date, loading cached meta information").length).to be == 2